+++ packages/fe-mockserver-core/src/data/dataAccess.ts
@@ -172,10 +172,9 @@ export class DataAccess implements DataAccessInterface {
             const actionName = odataRequest.queryPath[0].path;
             const entityContainerPath = this.metadata.getEntityContainerPath();
 
-            const actionDefinition: Action | undefined =
-                this.metadata.getVersion() === '2.0'
-                    ? this.metadata.getActionByFQN(`${entityContainerPath}/${actionName}`)
-                    : this.metadata.getActionImportByFQN(`${entityContainerPath}/${actionName}`)?.action;
+            const actionDefinition: Action | undefined = this.isV4()
+                ? this.metadata.getActionImportByFQN(`${entityContainerPath}/${actionName}`)?.action
+                : this.metadata.getActionByFQN(`${entityContainerPath}/${actionName}`);
 
             if (actionDefinition) {
                 if (actionDefinition.sourceType !== '') {
@@ -187,22 +186,14 @@ export class DataAccess implements DataAccessInterface {
                             odataRequest,
                             {}
                         );
-                        if (this.metadata.getVersion() === '2.0') {
+                        if (!this.isV4()) {
                             const enrichElement = (entitySet: EntitySet, dataLine: any) => {
-                                const keyValues: string[] = [];
-                                if (entitySet.entityType.keys.length === 1) {
-                                    keyValues.push(`${dataLine[entitySet.entityType.keys[0].name]}`);
-                                } else {
-                                    entitySet.entityType.keys.forEach((key) => {
-                                        keyValues.push(`${key.name}='${dataLine[key.name]}'`);
-                                    });
-                                }
-                                const uri = `${this.service.urlPath}/${entitySet.name}(${keyValues.join(',')})`;
-                                dataLine['__metadata'] = {
-                                    id: uri,
-                                    uri: uri,
-                                    type: entitySet.entityTypeName
-                                };
+                                const keyValues: Record<string, string> = {};
+                                entitySet.entityType.keys.forEach((key) => {
+                                    keyValues[key.name] = dataLine[key.name];
+                                });
+                                this.addV2Metadata(entitySet, keyValues, dataLine);
+
                                 return dataLine;
                             };
 
@@ -802,7 +793,7 @@ export class DataAccess implements DataAccessInterface {
                 data = data.slice(odataRequest.startIndex, odataRequest.startIndex + odataRequest.maxElements);
             }
 
-            if (this.metadata.getVersion() === '2.0') {
+            if (!this.isV4()) {
                 const enrichElement = (
                     entitySet: EntitySet | Singleton | undefined,
                     entityType: EntityType | undefined,
@@ -812,24 +803,14 @@ export class DataAccess implements DataAccessInterface {
                     if (!entityType) {
                         return;
                     }
-                    const keyValues: string[] = [];
-                    if (entityType.keys.length === 1) {
-                        const key = entityType.keys[0];
-                        keyValues.push(`${dataLine[key.name] || originalDataLine[key.name]}`);
-                    } else {
-                        entityType.keys.forEach((key) => {
-                            keyValues.push(`${key.name}='${dataLine[key.name] || originalDataLine[key.name]}'`);
-                        });
-                    }
+                    const keyValues: Record<string, string> = {};
+                    entityType.keys.forEach((key) => {
+                        keyValues[key.name] = dataLine[key.name] || originalDataLine[key.name];
+                    });
 
-                    const uri = `${this.service.urlPath}/${
-                        entitySet ? entitySet.name : entityType.name
-                    }(${keyValues.join(',')})`;
-                    dataLine['__metadata'] = {
-                        id: uri,
-                        uri: uri,
-                        type: entityType.fullyQualifiedName
-                    };
+                    if (entitySet) {
+                        this.addV2Metadata(entitySet, keyValues, dataLine);
+                    }
 
                     entityType.navigationProperties.forEach((navProp) => {
                         //eslint-disable-next-line
@@ -963,13 +944,12 @@ export class DataAccess implements DataAccessInterface {
                         .map((key) => `${key}='${currentKeys[key]}'`)
                         .join(',')})`
                 );
-                postData = (await this.getMockEntitySet(targetEntitySet.name)).performPOST(
-                    currentKeys,
-                    postData,
-                    odataRequest.tenantId,
-                    odataRequest,
-                    true
-                );
+                postData = await (
+                    await this.getMockEntitySet(targetEntitySet.name)
+                ).performPOST(currentKeys, postData, odataRequest.tenantId, odataRequest, true);
+                if (!this.isV4()) {
+                    this.addV2Metadata(entitySet, currentKeys, postData);
+                }
             } else {
                 if (!data[navPropertyName]) {
                     data[navPropertyName] = [];
@@ -985,13 +965,9 @@ export class DataAccess implements DataAccessInterface {
                     currentKeys[key.name] = postData[key.name];
                 }
             });
-            postData = (await this.getMockEntitySet(entitySet.name)).performPOST(
-                currentKeys,
-                postData,
-                odataRequest.tenantId,
-                odataRequest,
-                true
-            );
+            postData = await (
+                await this.getMockEntitySet(entitySet.name)
+            ).performPOST(currentKeys, postData, odataRequest.tenantId, odataRequest, true);
             odataRequest.setContext(`../$metadata#${entitySet.name}/$entity`);
             odataRequest.addResponseHeader(
                 'Location',
@@ -999,11 +975,31 @@ export class DataAccess implements DataAccessInterface {
                     .map((key) => `${key}='${currentKeys[key]}'`)
                     .join(',')})`
             );
+            if (!this.isV4()) {
+                this.addV2Metadata(entitySet, currentKeys, postData);
+            }
             odataRequest.setResponseData(await postData);
             return postData;
         } else {
             throw new Error('Unknown Entity Set' + entitySetName);
         }
+    }
+
+    private addV2Metadata(entitySet: EntitySet | Singleton, currentKeys: Record<string, string>, postData: any) {
+        let keyStr = '';
+        if (Object.keys(currentKeys).length === 1) {
+            keyStr = currentKeys[Object.keys(currentKeys)[0]];
+        } else {
+            keyStr = Object.keys(currentKeys)
+                .map((key) => `${key}='${currentKeys[key]}'`)
+                .join(',');
+        }
+        const uri = `${this.service.urlPath}/${entitySet.name}(${keyStr})`;
+        postData['__metadata'] = {
+            id: uri,
+            uri: uri,
+            type: entitySet.entityTypeName
+        };
     }
 
     public async deleteData(odataRequest: ODataRequest) {