+++ packages/fe-mockserver-core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/fe-mockserver-core
 
+## 1.1.48
+
+### Patch Changes
+
+-   Updated dependencies [b408038]
+    -   @sap-ux/annotation-converter@0.5.22
+
+## 1.1.47
+
+### Patch Changes
+
+-   fe1db88: Improve the result set for create on an odata v2 service
+
 ## 1.1.46
 
 ### Patch Changes